Added lost screen, improved pause and objects
Lost screen is now shown when the player looses.

Added AnimationManager where all animated objects are registered so that
they can be paused in a proper way.

Added GameObject, so that all game objects inherit from it and have a
common sprite property and are properly registered onto the
AnimationManager.