Cards are drawn to images before showing them.
CardManager does not own cards anymore, so that there may be more
than one instance of each card. When creating a card it will be
rendered to a texture, that all cards of the same type will share.

Since cardManager doesn't contain the cards, the loot choosing
needs to be modified.