Added support for story mode and ink integration
When creating the mainState, from now on MenuState will contain a static
member named gameConfiguration. This will be different depending on the
game mode chosen in the main menu.

The gameConfiguration will be in charge from now on of making
transitions across gameModes in the mainState.

Added support for overlay states to the mainState, this way the real
state will be paused but stop is not going to be called.

The endlessConfiguration makes the game behave as before.

The storyConfiguration loads the inky script and asks the newly added
dialogManager to show a line of dialog and its options.

DialogManager is a new game mode that will show the dialog. At the
moment it just prints the received dialog line.