Improved story handling and added messaging system.
Simplified configuration modes, removed getNextMode and getNextModeArguments. The update function receives the mainState, and can directly ask for a mode change to MainState.

StoryState now handles the game update through different story steps. When the ink document sends a GOTO command the story steps are changed.

Added a messaging system to the service locator. When the player fails a jump, it is now sent to this system and read both by the player class and by the jump tutorial story step.

The messaging system sends objects inheriting from the Message class. The different message types have different names, and objects can register for receiving only one of the message types.