Added escape mechanic and game state labels
When player runs out of sticks to throw it will have to escape.
This is accomplished by tapping quickly.

The top right corner of the screen will show the current state,
this way we let the player know what needs to be done. Some more
explanations will need to be added.

Also removed some energy related stuff from the player class.